refactor: move DonationReceipt to parent directory
