+++ src/Receipt/DonationReceipt.php
@@ -1,8 +1,7 @@
 <?php
-namespace Give\Receipt\DonationReceipt;
+namespace Give\Receipt;
 
 use Give\Helpers\ArrayDataSet;
-use Give\Receipt\Receipt;
 use function give_get_payment_meta as getDonationMetaData;
 use function give_get_gateway_admin_label as getGatewayLabel;
 use function give_get_donation_donor_email as getDonationDonorEmail;
